+++ Comms Over LAN/thermal_client.py
@@ -1,3 +1,4 @@
+""" NOT USED """
 import socket
 import time
